Move CIV creation methods to DisplaySet model
