+++ app/grandchallenge/reader_studies/views.py
@@ -14,7 +14,6 @@ from django.core.exceptions import (
 )
 from django.db import transaction
 from django.db.models import Count, Q
-from django.db.models.query import QuerySet
 from django.db.transaction import on_commit
 from django.forms import Form, Media
 from django.forms.utils import ErrorList
@@ -1406,14 +1405,13 @@ class DisplaySetUpdate(
         fields do not match the model fields: the model only has a `values`
         fields, whereas the form has a field for each value in those values.
         """
+        instance = self.get_object()
         for ci_slug, new_value in form.cleaned_data.items():
             if ci_slug == "order":
                 continue
-            self.create_civ(
-                ci_slug=ci_slug,
-                new_value=new_value,
+            instance.create_civ(
+                ci_slug=ci_slug, new_value=new_value, user=self.request.user
             )
-        instance = self.get_object()
         if (
             form.cleaned_data.get("order")
             and form.cleaned_data["order"] != instance.order
@@ -1422,75 +1420,6 @@ class DisplaySetUpdate(
             instance.save()
 
         return HttpResponseRedirect(self.get_success_url())
-
-    def create_civ_for_value(self, ci, current_civ, new_value):
-        instance = self.get_object()
-        current_value = current_civ.value if current_civ else None
-        if new_value and current_value != new_value:
-            instance.values.remove(current_civ)
-            civ = ComponentInterfaceValue.objects.create(
-                interface=ci, value=new_value
-            )
-            civ.full_clean()
-            instance.values.add(civ)
-        elif not new_value:
-            # if the new value is None, remove the old CIV from the display set
-            instance.values.remove(current_civ)
-
-    def create_civ_for_image(
-        self,
-        ci,
-        current_civ,
-        new_value,
-    ):
-        instance = self.get_object()
-        current_image = current_civ.image if current_civ else None
-        if isinstance(new_value, Image) and current_image != new_value:
-            instance.values.remove(current_civ)
-            civ, created = ComponentInterfaceValue.objects.get_or_create(
-                interface=ci, image=new_value
-            )
-            if created:
-                civ.full_clean()
-            instance.values.add(civ)
-        elif isinstance(new_value, QuerySet):
-            us = RawImageUploadSession.objects.create(
-                creator=self.request.user,
-            )
-            us.user_uploads.set(new_value)
-            us.process_images(
-                linked_task=add_image_to_display_set.signature(
-                    kwargs={
-                        "display_set_pk": instance.pk,
-                        "interface_pk": str(ci.pk),
-                    },
-                    immutable=True,
-                )
-            )
-
-    def create_civ_for_file(self, ci, current_civ, new_value):
-        instance = self.get_object()
-        # in this case, new_value is an instance of a CIV already (or None)
-        if new_value and current_civ != new_value:
-            instance.values.remove(current_civ)
-            instance.values.add(new_value)
-        elif not new_value:
-            # if no new value is provided (user selects '---' in dropdown)
-            # delete old CIV
-            instance.values.remove(current_civ)
-
-    def create_civ(self, ci_slug, new_value):
-        instance = self.get_object()
-        ci = ComponentInterface.objects.get(slug=ci_slug)
-        current_civ = instance.values.filter(interface=ci).first()
-        if ci.is_json_kind and not ci.requires_file:
-            return self.create_civ_for_value(ci, current_civ, new_value)
-        elif ci.is_image_kind:
-            return self.create_civ_for_image(ci, current_civ, new_value)
-        elif ci.requires_file:
-            return self.create_civ_for_file(ci, current_civ, new_value)
-        else:
-            NotImplementedError(f"CIV creation for {ci} not handled.")
 
 
 class DisplaySetFilesUpdate(ObjectPermissionRequiredMixin, FormView):
